+++ src/EventListener/LoginSuccessListener.php
@@ -43,6 +43,7 @@ class LoginSuccessListener implements EventSubscriberInterface
             $eventMetadata = [
                 'platform' => $platformMode,
                 'ip' => $_SERVER['REMOTE_ADDR'],
+                'user_agent' => $_SERVER['HTTP_USER_AGENT'] ?? 'Unknown',
                 'uuid' => $user->getUuid(),
             ];
             $this->eventActions->saveEvent(